Fix config syntax
